+++ api/trace/testtrace/span_test.go
@@ -33,12 +33,13 @@ import (
 
 func TestSpan(t *testing.T) {
 	t.Run("#Tracer", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("returns the tracer used to start the span", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, subject := tracer.Start(context.Background(), "test")
 
 			e.Expect(subject.Tracer()).ToEqual(tracer)
@@ -46,12 +47,13 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#End", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("ends the span", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -82,7 +84,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -105,7 +107,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -124,6 +126,7 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#RecordError", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("records an error", func(t *testing.T) {
 			t.Parallel()
 
@@ -147,7 +150,7 @@ func TestSpan(t *testing.T) {
 			for _, s := range scenarios {
 				e := matchers.NewExpecter(t)
 
-				tracer := testtrace.NewTracer()
+				tracer := tp.Tracer(t.Name())
 				ctx, span := tracer.Start(context.Background(), "test")
 
 				subject, ok := span.(*testtrace.Span)
@@ -175,7 +178,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			ctx, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -204,7 +207,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			ctx, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -221,7 +224,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			ctx, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -234,12 +237,13 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#IsRecording", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("returns true", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, subject := tracer.Start(context.Background(), "test")
 
 			e.Expect(subject.IsRecording()).ToBeTrue()
@@ -247,12 +251,13 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#SpanContext", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("returns a valid SpanContext", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, subject := tracer.Start(context.Background(), "test")
 
 			e.Expect(subject.SpanContext().IsValid()).ToBeTrue()
@@ -263,7 +268,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, subject := tracer.Start(context.Background(), "test")
 
 			e.Expect(subject.SpanContext()).ToEqual(subject.SpanContext())
@@ -271,12 +276,13 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Name", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("returns the most recently set name on the span", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			originalName := "test"
 			_, span := tracer.Start(context.Background(), originalName)
 
@@ -299,7 +305,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			originalName := "test"
 			_, span := tracer.Start(context.Background(), originalName)
 
@@ -314,12 +320,13 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Attributes", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("returns an empty map by default", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -333,7 +340,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -359,7 +366,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -382,7 +389,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -409,12 +416,13 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Links", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("returns an empty map by default", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -425,12 +433,13 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Events", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("returns an empty slice by default", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -444,7 +453,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -497,7 +506,7 @@ func TestSpan(t *testing.T) {
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -518,12 +527,13 @@ func TestSpan(t *testing.T) {
 	})
 
 	t.Run("#Status", func(t *testing.T) {
+		tp := testtrace.NewProvider()
 		t.Run("defaults to OK", func(t *testing.T) {
 			t.Parallel()
 
 			e := matchers.NewExpecter(t)
 
-			tracer := testtrace.NewTracer()
+			tracer := tp.Tracer(t.Name())
 			_, span := tracer.Start(context.Background(), "test")
 
 			subject, ok := span.(*testtrace.Span)
@@ -562,7 +572,7 @@ func TestSpan(t *testing.T) {
 
 				e := matchers.NewExpecter(t)
 
-				tracer := testtrace.NewTracer()
+				tracer := tp.Tracer(t.Name())
 				_, span := tracer.Start(context.Background(), "test")
 
 				subject, ok := span.(*testtrace.Span)
@@ -580,7 +590,7 @@ func TestSpan(t *testing.T) {
 
 				e := matchers.NewExpecter(t)
 
-				tracer := testtrace.NewTracer()
+				tracer := tp.Tracer(t.Name())
 				_, span := tracer.Start(context.Background(), "test")
 
 				subject, ok := span.(*testtrace.Span)